requirements.txt
apturl==0.5.2
asgiref==3.4.1
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
bcrypt==3.1.7
bleach==4.1.0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.9
django-allauth==0.47.0
django-countries==7.2.1
django-crispy-forms==1.13.0
django-dynamic-filenames==1.2.0
django-resized==0.3.11
django-stdimage==5.3.0
django-summernote==0.8.20.0
django-tinymce==3.4.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
gpg===1.13.1-unknown
httplib2==0.14.0
idna==2.8
ipdb==0.13.9
ipython==7.30.1
jedi==0.18.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib-inline==0.1.3
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
piglet==1.0.0
piglet-templates==1.2.0
Pillow==7.0.0
prompt-toolkit==3.0.23
protobuf==3.6.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pyfiglet==0.7
Pygments==2.10.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.6
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
systemd-python==234
toml==0.10.2
traitlets==5.1.1
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unicode-slugify==0.1.5
Unidecode==1.3.2
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
xkit==0.0.0
